requirements.txt
asgi-redis==1.4.3
asgiref==1.1.2
async-timeout==2.0.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
certifi==2018.1.18
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
daphne==1.3.0
Django==1.11.6
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
msgpack-python==0.4.7
oauthlib==2.0.6
olefile==0.44
Pillow==4.3.0
pkg-resources==0.0.0
pytz==2017.3
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
urllib3==1.22
websocket-client==0.46.0
zope.interface==4.4.3
